requirements.txt
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-google-genai==3.0.0
faiss-cpu==1.12.0
pydantic==2.12.3
python-dotenv==1.0.1
streamlit==1.32.0
rich==13.7.0
